requirements.txt
# requirements.txt
langgraph 
langchain-ollama
langchain-community
langchain-deepseek-official
langchain
pydantic
faiss-cpu
chromadb
